replaced jdbcTemplate by querydsl moved configuration from xml to java replaced c3p0 by HikariCP replaced commons-logging by slf4j made startDate and endDate optional improvements

+++ src/main/java/org/bitbucket/eluinstra/fs/core/file/FSFileDAO.java
@@ -23,6 +23,6 @@ public interface FSFileDAO
 {
 	boolean isAuthorized(@NonNull byte[] certificate, @NonNull String path);
 	Optional<FSFile> findFileByVirtualPath(@NonNull String path);
-	int insertFile(@NonNull FSFile fsFile);
-	int deleteFile(@NonNull String path);
+	long insertFile(@NonNull FSFile fsFile);
+	long deleteFile(@NonNull String path);
 }
